All PR live tests reside in the same class
