+++ src/test/java/com/cdancy/bitbucket/rest/features/PullRequestApiLiveTest.java
@@ -70,8 +70,7 @@ public class PullRequestApiLiveTest extends BaseBitbucketApiLiveTest {
     private int prId = -1;
     private int version = -1;
 
-    @BeforeClass
-    public void init() {
+    private void setupGit() {
         generatedTestContents = TestUtilities.initGeneratedTestContents(this.endpoint, this.bitbucketAuthentication, this.api);
         this.project = generatedTestContents.project.key();
         this.repo = generatedTestContents.repository.name();
@@ -89,13 +88,16 @@ public class PullRequestApiLiveTest extends BaseBitbucketApiLiveTest {
         }
 
         assertThat(branchToMerge).isNotNull();
+    }
 
+    @BeforeClass
+    public void init() {
+        setupGit();
         addTestUser();
         testUserApi = apiForTestUser();
     }
 
-    @Test
-    public void createPullRequest() {
+    private void createPullRequest() {
         final String randomChars = TestUtilities.randomString();
         final ProjectKey proj = ProjectKey.create(project);
         final MinimalRepository repository = MinimalRepository.create(repo, null, proj);
@@ -110,7 +112,12 @@ public class PullRequestApiLiveTest extends BaseBitbucketApiLiveTest {
         version = pr.version();
     }
 
-    @Test (dependsOnMethods = "createPullRequest")
+    @Test
+    public void testCreatePullRequest() {
+        createPullRequest();
+    }
+
+    @Test (dependsOnMethods = "testCreatePullRequest")
     public void testGetPullRequest() {
         final PullRequest pr = api().get(project, repo, prId);
         assertThat(pr).isNotNull();
@@ -287,11 +294,48 @@ public class PullRequestApiLiveTest extends BaseBitbucketApiLiveTest {
         assertThat(success.errors()).isNotEmpty();
     }
 
+    @Test (dependsOnMethods = "testDeleteMergedPullRequest")
+    public void testDeleteBadVersionOfPullRequest() {
+        // Since the PR has been merged by another test, we now need a new PR.
+        // To get a new PR, we need to create new branch in git, push it and open a new PR with it.
+        terminateGitContents();
+        setupGit();
+        createPullRequest();
+        final DeletePullRequest deletePullRequest = DeletePullRequest.create(9999);
+        final RequestStatus success = api().delete(project, repo, prId, deletePullRequest);
+        assertThat(success).isNotNull();
+        assertThat(success.value()).isFalse();
+        assertThat(success.errors()).isNotEmpty();
+    }
+
+    @Test (dependsOnMethods = "testDeleteBadVersionOfPullRequest")
+    public void testDeleteNonExistentPullRequest() {
+        final DeletePullRequest deletePullRequest = DeletePullRequest.create(version);
+        final RequestStatus success = api().delete(project, repo, 9999, deletePullRequest);
+        assertThat(success).isNotNull();
+        assertThat(success.value()).isFalse();
+        assertThat(success.errors()).isNotEmpty();
+    }
+
+    @Test (dependsOnMethods = "testDeleteNonExistentPullRequest")
+    public void testDeletePullRequest() {
+        final DeletePullRequest deletePullRequest = DeletePullRequest.create(version);
+        final RequestStatus success = api().delete(project, repo, prId, deletePullRequest);
+        assertThat(success).isNotNull();
+        assertThat(success.value()).isTrue();
+        assertThat(success.errors()).isEmpty();
+    }
+
+
     @AfterClass
     public void fin() throws IOException {
-        TestUtilities.terminateGeneratedTestContents(this.api, generatedTestContents);
+        terminateGitContents();
         testUserApi.close();
         deleteTestUser();
+    }
+
+    private void terminateGitContents() {
+        TestUtilities.terminateGeneratedTestContents(this.api, generatedTestContents);
     }
 
     private PullRequestApi api() {
